+++ CHANGELOG.md
@@ -618,3 +618,4 @@
 - [2026-04-23 21:09] test(theme): implement widget integration tests for theme
 - [2026-04-23 21:09] style(auth): harmonize light and dark theme palette for auth
 - [2026-04-23 21:10] docs(core): update readme and API contracts for core
+- [2026-04-23 21:11] docs(network): update readme and API contracts for network